pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "pingrunhuang", email = "pingrunhuang@gmail.com"},
]
dependencies = [
    "jupyter>=1.0.0",
    "pip>=23.1.2",
    "tensorflow>=2.11.1",
    "tensorflow-io-gcs-filesystem==0.27.0",
    "pymongo>=4.4.0",
    "motor>=3.2.0",
    "pandas>=2.0.2",
    "matplotlib>=3.7.1",
    "scikit-learn>=1.2.2",
    "numpy>=1.25.0",
    "PyYaml>=6.0",
]
requires-python = ">=3.9,<3.11"
license = {text = "MIT"}
readme = "README.md"
